add default values for vm-monitor arguments
Currently, the AUTOSCALING environment variable is being set by the
control plane. However, it does not send arguments yet. Thus, when this
monitor is merged into compute_ctl, at first it will think it's
AUTOSCALING, but it won't be provided any arguments. We need defaults,
otherwise the monitor will panic with "unprovided arguments" or
something of the like. Once cloud/6331 (add vm-monitor) arguments is
merged, these defaults can be removed.

These defaults are the arguments used to connect to the vm-informant.
The arguments for connecting to an agent are slightly different.
Instead, the monitor should listen on 0.0.0.0:10301 in that case.